% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateOptimizationParameters.R
\name{buildMomentAndDerivativesFunctions}
\alias{buildMomentAndDerivativesFunctions}
\title{buildMomentAndDerivativesFunctions}
\usage{
buildMomentAndDerivativesFunctions(a, m = NULL, d = NULL, ...)
}
\arguments{
\item{d}{}
}
\description{
buildMomentAndDerivativesFunctions
}
\examples{
a <- 0
m <- 0:2
d <- 1:3

# Create the functions to estimate the derivatives and moments
MyFunc <- GLP:::buildMomentAndDerivativesFunctions(a,m,d)

# Create a sample
sample <- rnorm(1E4, 0, 1 )
Truth <- c((a^2 - 1)*dnorm(a), -a*dnorm(a), dnorm(a), 1- pnorm(a), (2*pi)^(-1/2), 1/2)
data.frame(Truth = Truth , Estimates = sapply(MyFunc, function(f)do.call(f, list(x = sample))))
}

