% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateOptimizationParameters.R
\name{cleanAndcheckdAndm}
\alias{cleanAndcheckdAndm}
\title{Title}
\usage{
cleanAndcheckdAndm(d, m)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
Title
}
\examples{
d <- c(4,2,1)
m <- 0:3
cleanAndcheckdAndm(d,m)

d <- c(1/2,2,1)
cleanAndcheckdAndm(d,m)

d <- 0:3
m <- 1:3
cleanAndcheckdAndm(d,m)


}

