% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateOptimizationParameters.R
\name{getBootstrapedValues}
\alias{getBootstrapedValues}
\title{Title}
\usage{
getBootstrapedValues(sample, fboot, nboot = 1000, mc.cores = 1)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
sample <- rnorm(1e4,0,1)
fboot <-  list(function(x) mean(x), function(x) sd(x))
bootstrapedValues <- getBootstrapedValues(sample,fboot, nboot = 1E4, mc.cores = 4)
head(bootstrapedValues)

}

