% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateOptimizationParameters.R
\name{plotCI}
\alias{plotCI}
\title{Plot hyperrectangular and ellipsoidal confidence intervals}
\usage{
plotCI(sample, CI)
}
\arguments{
\item{sample}{Vector of sample points}

\item{CI}{List containing either \emph{hyperrectangle}{, data.frame containing the lower bounds
and upper bounds of the derivatives and truncated moments.}, or \item{ellipsoid}{, a list containing the vector of means,
the covariance matrix, and the radius of the ellipsoid describing the confidence interval}}
}
\value{
an object of class ggplot2
}
\description{
Plot hyperrectangular and ellipsoidal confidence intervals
}
\examples{
set.seed(100)
n <- 300 ;  mu <- c(1,2,3) ; Sigma <- matrix(c(6,0,0,0,1,1,0,1,1),ncol =3) ; alpha = 0.05
sample <- MASS::mvrnorm(n,mu,Sigma) \%>\% as.data.frame
CIellipsoid <- list(mu = colMeans(sample, na.rm = TRUE), Sigma = cov(sample, use = "complete.ob"), radius = sqrt(qchisq(1-alpha/2, df = 2)))
CIrectangular <- apply(sample, 2, quantile, probs = c(alpha/2,1-alpha/2))
CI <- list(ellipsoid = CIellipsoid, hyperrectangle = CIrectangular)
plotCI(sample,CI)

n <- 300 ;  mu <- c(1,2) ; Sigma <- matrix(c(6,1,1,1),ncol =2) ; alpha = 0.05
sample <- MASS::mvrnorm(n,mu,Sigma) \%>\% as.data.frame
CIellipsoid <- list(mu = colMeans(sample, na.rm = TRUE), Sigma = cov(sample, use = "complete.ob"), radius = sqrt(qchisq(1-alpha/2, df = 2)))
CIrectangular <- apply(sample, 2, quantile, probs = c(alpha/2,1-alpha/2))
CI <- list(ellipsoid = CIellipsoid, hyperrectangle = CIrectangular)
plotCI(sample,CI)
}

