% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{computeBound}
\alias{computeBound}
\title{Title}
\usage{
computeBound(H, mu, sigma, lambda, scale = 1, direction = "max")
}
\arguments{
\item{direction}{}
}
\description{
Title
}
\examples{
###
### Solve problem with equalities
###
# Function and parameters for the integral of the objective
H <- function(x) as.numeric(qexp(0.9) <= x)
mu <- 1
sigma <- 2
lambda <- 0

output <- computeBound(H,mu, sigma, lambda)

# Check that the bound matches the analytical solution in Bertsimas
CMsquare <- (sigma- mu^2)/mu^2
delta <-  qexp(0.9)/mu-1

data.frame(Algorithm = output$bound, Analytical = CMsquare/(CMsquare + delta^2))

# Check that the output is feasible
with(output$P, data.frame(moments = c(sum(p),sum(p*x), sum(p*x^2)), truth = c(1,mu,sigma) ))

###
### Solve problem with equalities
###
mu <- c(1,2)
sigma <- c(2,4)
lambda <- 0

output <- computeBound(H,mu, sigma, lambda)

# Check that the bound is larger (because we increase the feasible set) than the analytical solution in Bertsimas
data.frame(Algorithm = output$bound, Analytical = CMsquare/(CMsquare + delta^2))

# Check that the output is feasible
with(output$P, data.frame(lowerMomentBound = c(1, mu[1], sigma[1]), moments = c(sum(p),sum(p*x), sum(p*x^2)), upperMomentBound = c(1, mu[2], sigma[2]) ))
}

