% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{computeBoundVal}
\alias{computeBoundVal}
\title{Title}
\usage{
computeBoundVal(H, mu, sigma, lambda, direction = "max")
}
\arguments{
\item{direction}{}
}
\description{
Title
}
\examples{
####
#### Finding a the optimal probability measure (p,x) to the problem
####
#### max P(X > c)
#### s.t. sum(p)  = 1
####      sum(px) = 1
####      sum(px^2) = 2
####
#### where c is the 90th percentile of a standard exponential distribution
#### Note that the solution to this problem is known (see Theorem 3.3 of Bertsimas and Popescu)

# Function and parameters for the integral of the objective
H <- function(x) as.numeric(qexp(0.9) <= x)
mu <- 1
sigma <- 2
lambda <- 0

output <- computeBoundVal(H,mu, sigma, lambda)

# Check that the bound matches the analytical solution in Bertsimas
CMsquare <- (sigma- mu^2)/mu^2
delta <-  qexp(0.9)/mu-1

data.frame(Algorithm = output$bound, Analytical = CMsquare/(CMsquare + delta^2))

# Check that the output is feasible
with(output$P, data.frame(moments = c(sum(p),sum(p*x), sum(p*x^2)), truth = c(1,mu,sigma) ))
}

