% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateOptimizationParameters.R
\name{getCIMomentAndDerivatives}
\alias{getCIMomentAndDerivatives}
\title{Title}
\usage{
getCIMomentAndDerivatives(sample, a, m = NULL, d = NULL, nboot = 1000,
  alpha = 0.05, method = c("hyperrectangle", "ellipsoid", "both"),
  mc.cores = 1, bootSample = FALSE)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
a <- c(0,1) ; m <- c(0, 1) ; d <- 1 ; nboot <- 1000
set.seed(100) ; sample <- rnorm(100, 0, 1)
hist(sample)
CI <- getCIMomentAndDerivatives(sample, a,m,d,nboot = nboot,mc.cores = 4, method = "both", bootSample = TRUE)

i <- 2
title <- paste0("95\%-CI of the Estimated Parameters when a = ",a[i])
plot <- plotCI(CI[[i]]$bootSample, CI[[i]])
plot + geom_point(aes(x = dnorm(a[i]), y = 1-pnorm(a[i])), colour = "blue") +
geom_text(aes(x = 1.009*dnorm(a[i]), y = 1-pnorm(a[i]), label = "True Value"), colour = "blue")+
labs(x = "Derivative of order 1", y = "Moment of order 0") + ggtitle(title)

}

